requirements.txt
PyDriller==2.6
pytz==2023.3.post1
pandas==2.1.4
PyYAML==6.0.1
openai==0.28.0
javalang==0.13.0
GitPython==3.1.40
tiktoken==0.5.2
pinecone-client==3.0.0

numpy~=1.26.2
setuptools~=65.5.0
requests~=2.31.0
tenacity~=8.2.3
urllib3~=2.1.0
six~=1.16.0
matplotlib~=3.8.2
filelock~=3.13.1
langchain~=0.1.1
tqdm~=4.66.1